Implemented letting a user change their password
Implement PUT for the login/ endpoint; this allows a user to change
their password by PUTing a resource with the new password to login/.
Note that they will need to confirm that they recieved a OK response, or
they may lock themselves out of their own account. I also implemented
some tests for this functionality.